requirements.txt
prettytable==0.7
termcolor==1.1.0
python-bugzilla>=0.7.0
stevedore
oslo.config
